Change `AccountOAuthScreen()` to `AccountOAuthView()`
