+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/ui/view/SignInView.kt
@@ -10,14 +10,11 @@ import androidx.compose.ui.text.style.TextAlign
 import app.k9mail.core.ui.compose.common.DevicePreviews
 import app.k9mail.core.ui.compose.designsystem.atom.button.Button
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextCaption
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextSubtitle1
-import app.k9mail.core.ui.compose.designsystem.molecule.input.EmailAddressInput
 import app.k9mail.core.ui.compose.theme.MainTheme
 import app.k9mail.feature.account.oauth.R
 
 @Composable
 internal fun SignInView(
-    emailAddress: String,
     onSignInClick: () -> Unit,
     isGoogleSignIn: Boolean,
     modifier: Modifier = Modifier,
@@ -27,14 +24,6 @@ internal fun SignInView(
         verticalArrangement = Arrangement.spacedBy(MainTheme.spacings.double),
         modifier = modifier,
     ) {
-        TextSubtitle1(text = stringResource(id = R.string.account_oauth_sign_in_title))
-
-        EmailAddressInput(
-            emailAddress = emailAddress,
-            onEmailAddressChange = {},
-            isEnabled = false,
-        )
-
         TextCaption(
             text = stringResource(id = R.string.account_oauth_sign_in_description),
             textAlign = TextAlign.Center,
@@ -57,7 +46,6 @@ internal fun SignInView(
 @Composable
 internal fun SignInViewPreview() {
     SignInView(
-        emailAddress = "test@example.com",
         onSignInClick = {},
         isGoogleSignIn = false,
     )
@@ -67,7 +55,6 @@ internal fun SignInViewPreview() {
 @Composable
 internal fun SignInViewWithGooglePreview() {
     SignInView(
-        emailAddress = "test@gmail.com",
         onSignInClick = {},
         isGoogleSignIn = true,
     )
